chat/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.5
cssselect==1.1.0
Django==4.0.1
django-channels==0.7.0
djangorestframework==3.13.1
fake-useragent==0.1.11
feedparser==6.0.10
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.1
multitasking==0.0.11
numpy==1.23.1
oauthlib==3.2.0
pandas==1.4.3
parse==1.19.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2021.5
urllib3==1.26.8
w3lib==1.22.0
websockets==10.3
yahoo-fin==0.8.9.1
yfinance==0.1.74
zipp==3.8.1
